backend/requirements.txt
fastapi
uvicorn 
requests 
httpx
rapidfuzz
fpdf
google-cloud-storage
google-auth
asyncpg
pandas